requirements.txt
pyenchant
contractions
pre-commit
